# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Overtime App",
        "description": "This is a sample server overtime api",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/overtime-pays/calculate": {
            "get": {
                "tags": [
                    "Overtime Pays"
                ],
                "summary": "Get calculate of overtime pays",
                "description": "Returns list of overtime pays calculate",
                "operationId": "getOvertimePaysCalculate",
                "parameters": [
                    {
                        "name": "month",
                        "in": "query",
                        "required": true,
                        "example": "2022-05"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "meta": {
                                            "type": "object"
                                        },
                                        "data": {
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    }
}